+++ Sources/BaseFoundationiOS/Templates/Structs/States/FormState.swift
@@ -1,0 +1,8 @@
+//
+//  File.swift
+//  
+//
+//  Created by pablo borquez on 20/12/23.
+//
+
+import Foundation